refactor: move struct and utils to jf_helpers.rs
