+++ Cargo.toml
@@ -5,6 +5,7 @@ resolver = "2"
 members = ["contract-bindings", "contracts/rust", "sequencer", "utils"]
 
 [workspace.dependencies]
+hotshot-state-prover = { git = "https://github.com/EspressoSystems/HotShot.git", rev = "0dde152df0af0c799f906ea14a2b59ade98b6d18" }
 jf-plonk = { git = "https://github.com/EspressoSystems/jellyfish", features = [
   "test-apis",
   "test-srs",
