+++ contracts/rust/src/lib.rs
@@ -1,3 +1,4 @@
 mod bls_contract;
 mod helpers;
 mod hotshot_contract;
+pub mod jf_helpers;
